# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeArena competitive programming platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export const AckResponseValue = {
  received: true,
} as const;
export type AckResponse = typeof AckResponseValue;

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface Language {
  id: number;
  name: string;
  judge0Id: number;
  monacoId: string;
  timeMultiplier: number;
  memoryMultiplier: number;
}

export type ProblemDifficulty = typeof ProblemDifficulty[keyof typeof ProblemDifficulty];


export const ProblemDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface ProblemSummary {
  id: number;
  slug: string;
  title: string;
  difficulty: ProblemDifficulty;
  points: number;
  tags: string[];
  solvedCount: number;
  createdAt: string;
}

export interface TestCase {
  id: number;
  problemId: number;
  stdin: string;
  expectedOutput: string;
  isSample: boolean;
  points: number;
  orderIndex: number;
}

export interface Problem {
  id: number;
  slug: string;
  title: string;
  description: string;
  difficulty: ProblemDifficulty;
  points: number;
  tags: string[];
  /** Base CPU time limit in seconds, multiplied per-language. */
  cpuTimeLimitSeconds: number;
  /** Base memory limit in kilobytes, multiplied per-language. */
  memoryLimitKb: number;
  sampleTestCases: TestCase[];
  solvedCount: number;
  createdAt: string;
}

export interface ProblemInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  difficulty: ProblemDifficulty;
  /** @minimum 0 */
  points: number;
  tags?: string[];
  /** @minimum 0.1 */
  cpuTimeLimitSeconds: number;
  /** @minimum 1024 */
  memoryLimitKb: number;
}

export interface ProblemUpdate {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  description?: string;
  difficulty?: ProblemDifficulty;
  /** @minimum 0 */
  points?: number;
  tags?: string[];
  /** @minimum 0.1 */
  cpuTimeLimitSeconds?: number;
  /** @minimum 1024 */
  memoryLimitKb?: number;
}

export interface TestCaseInput {
  stdin: string;
  expectedOutput: string;
  isSample: boolean;
  /** @minimum 0 */
  points: number;
}

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  queued: 'queued',
  judging: 'judging',
  accepted: 'accepted',
  partial: 'partial',
  wrong_answer: 'wrong_answer',
  time_limit_exceeded: 'time_limit_exceeded',
  memory_limit_exceeded: 'memory_limit_exceeded',
  runtime_error: 'runtime_error',
  compilation_error: 'compilation_error',
  internal_error: 'internal_error',
} as const;

export type TestResultStatus = typeof TestResultStatus[keyof typeof TestResultStatus];


export const TestResultStatus = {
  queued: 'queued',
  processing: 'processing',
  accepted: 'accepted',
  wrong_answer: 'wrong_answer',
  time_limit_exceeded: 'time_limit_exceeded',
  memory_limit_exceeded: 'memory_limit_exceeded',
  runtime_error: 'runtime_error',
  compilation_error: 'compilation_error',
  internal_error: 'internal_error',
} as const;

export interface SubmissionTestResult {
  id: number;
  testCaseId: number;
  status: TestResultStatus;
  isSample: boolean;
  points: number;
  earnedPoints: number;
  /** @nullable */
  timeMs?: number | null;
  /** @nullable */
  memoryKb?: number | null;
  /** @nullable */
  stdout?: string | null;
  /** @nullable */
  stderr?: string | null;
  /** @nullable */
  compileOutput?: string | null;
}

export interface SubmissionInput {
  problemId: number;
  /** @nullable */
  contestId?: number | null;
  languageId: number;
  /** @minLength 1 */
  sourceCode: string;
}

export interface SubmissionSummary {
  id: number;
  problemId: number;
  problemTitle: string;
  /** @nullable */
  contestId?: number | null;
  languageId: number;
  languageName: string;
  status: SubmissionStatus;
  score: number;
  maxScore: number;
  submittedAt: string;
}

export interface Submission {
  id: number;
  problemId: number;
  problemTitle: string;
  /** @nullable */
  contestId?: number | null;
  languageId: number;
  languageName: string;
  sourceCode: string;
  status: SubmissionStatus;
  score: number;
  maxScore: number;
  submittedAt: string;
  /** @nullable */
  judgedAt?: string | null;
  testResults: SubmissionTestResult[];
}

export type Judge0CallbackPayloadStatus = {
  id?: number;
  description?: string;
};

export interface Judge0CallbackPayload {
  token: string;
  /** @nullable */
  stdout?: string | null;
  /** @nullable */
  stderr?: string | null;
  /** @nullable */
  compile_output?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  time?: string | null;
  /** @nullable */
  memory?: number | null;
  status?: Judge0CallbackPayloadStatus;
}

export interface ContestProblem {
  problemId: number;
  title: string;
  difficulty: ProblemDifficulty;
  points: number;
  label: string;
}

export interface ContestSummary {
  id: number;
  slug: string;
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  problemCount: number;
}

export interface Contest {
  id: number;
  slug: string;
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  problems: ContestProblem[];
}

export interface ContestProblemInput {
  problemId: number;
  /** @minimum 0 */
  points: number;
  label: string;
}

export interface ContestInput {
  /** @minLength 1 */
  title: string;
  description: string;
  startTime: string;
  endTime: string;
  problems: ContestProblemInput[];
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  /** @nullable */
  profileImageUrl: string | null;
  totalScore: number;
  solvedCount: number;
  penaltyMinutes: number;
  /** @nullable */
  lastAcceptedAt: string | null;
}

export interface DashboardSummary {
  solvedCount: number;
  totalSubmissions: number;
  acceptanceRate: number;
  recentSubmissions: SubmissionSummary[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type ListProblemsParams = {
difficulty?: ProblemDifficulty;
};

export type ListSubmissionsParams = {
problemId?: number;
contestId?: number;
};