# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeArena competitive programming platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /judge0/callback/{testResultId}:
    post:
      operationId: receiveJudge0Callback
      tags: [judge0]
      summary: Internal webhook invoked by Judge0 when a test case finishes executing
      parameters:
        - name: testResultId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Judge0CallbackPayload"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AckResponse"
  /contests:
    get:
      operationId: listContests
      tags: [contests]
      summary: List contests
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContestSummary"
    post:
      operationId: createContest
      tags: [contests]
      summary: Create a contest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contests/{id}:
    get:
      operationId: getContest
      tags: [contests]
      summary: Get a contest with its problem set
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contests/{id}/leaderboard:
    get:
      operationId: getContestLeaderboard
      tags: [contests]
      summary: Get the live standings for a contest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    AckResponse:
      type: object
      required: [received]
      properties:
        received:
          type: boolean
          const: true
...
    ProblemDifficulty:
      type: string
      enum: [easy, medium, hard]
...
    Judge0CallbackPayload:
      type: object
      required: [token]
      properties:
        token:
          type: string
        stdout:
          type: [string, 'null']
        stderr:
          type: [string, 'null']
        compile_output:
          type: [string, 'null']
        message:
          type: [string, 'null']
        time:
          type: [string, 'null']
        memory:
          type: [number, 'null']
        status:
          type: object
          properties:
            id:
              type: integer
            description:
              type: string
    ContestProblem:
      type: object
      required: [problemId, title, difficulty, points, label]
      properties:
        problemId:
          type: integer
        title:
          type: string
        difficulty:
          $ref: "#/components/schemas/ProblemDifficulty"
        points:
          type: integer
        label:
          type: string
    ContestSummary:
      type: object
      required: [id, slug, title, description, startTime, endTime, problemCount]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        problemCount:
          type: integer
    Contest:
      type: object
      required: [id, slug, title, description, startTime, endTime, problems]
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ContestProblem"
    ContestProblemInput:
      type: object
      required: [problemId, points, label]
      properties:
        problemId:
          type: integer
        points:
          type: integer
          minimum: 0
        label:
          type: string
    ContestInput:
      type: object
      required: [title, description, startTime, endTime, problems]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ContestProblemInput"
    LeaderboardEntry:
      type: object
      required: [rank, userId, username, profileImageUrl, totalScore, solvedCount, penaltyMinutes, lastAcceptedAt]
      properties:
        rank:
          type: integer
        userId:
          type: string
        username:
          type: string
        profileImageUrl:
          type: [string, 'null']
        totalScore:
          type: integer
        solvedCount:
          type: integer
        penaltyMinutes:
          type: integer
        lastAcceptedAt:
          type: [string, 'null']
          format: date-time
...