# lib/api-zod/src/generated/types/contestProblem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeArena competitive programming platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDifficulty } from './problemDifficulty';

export interface ContestProblem {
  problemId: number;
  title: string;
  difficulty: ProblemDifficulty;
  points: number;
  label: string;
}