# lib/api-zod/src/generated/types/contestProblemInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * CodeArena competitive programming platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContestProblemInput {
  problemId: number;
  /** @minimum 0 */
  points: number;
  label: string;
}